requirements.txt
django==4.1.5
djangorestframework==3.14.0
markdown==3.4.1
django-filter==22.1
psycopg==3.1.7
pillow==9.4.0
djangorestframework-simplejwt==5.2.2
drf-writable-nested==0.7.0
django-cors-headers==3.13.0
zeep==4.2.1
aiohttp==3.8.3